+++ WorldCities/WorldCities/src/app/app.component.ts
@@ -1,6 +1,5 @@
-import { HttpClient } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
-
+import { AuthService } from './auth/auth.service';
 
 @Component({
   selector: 'app-root',
@@ -8,10 +7,13 @@ import { Component, OnInit } from '@angular/core';
   styleUrl: './app.component.scss'
 })
 export class AppComponent implements OnInit {
+
   title = 'WorldCities';
 
+  constructor(private authService: AuthService) { }
+
   ngOnInit(): void {
- 
+    this.authService.init();
   }
 
 }