+++ WorldCities/WorldCities/src/app/auth/auth.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs';
+import { Observable, Subject, tap } from 'rxjs';
 
 import { environment } from './../../envrionments/environment';
 import { LoginRequest } from './login-request';
@@ -12,7 +12,9 @@ import { LoginResult } from './login-result';
 })
 export class AuthService {
 
-  public tokenKey: string = "token";
+  private tokenKey: string = "token";
+  private _authStatus = new Subject<boolean>();
+  public authStatus = this._authStatus.asObservable();
 
   constructor(protected http: HttpClient) { }
 
@@ -25,9 +27,29 @@ export class AuthService {
     return localStorage.getItem(this.tokenKey);
   }
 
+  init(): void {
+    if (this.isAuthenticated()) this.setAuthStatus(true);
+  }
+
   login(item: LoginRequest): Observable<LoginResult> {
     var url = environment.baseURL + "api/Account/Login";
-    return this.http.post<LoginResult>(url, item);
+    return this.http
+      .post<LoginResult>(url, item)
+      .pipe(tap(loginResult => {
+        if (loginResult.success && loginResult.token) {
+          localStorage.setItem(this.tokenKey, loginResult.token);
+          this.setAuthStatus(true);
+        }
+      }));
+  }
+
+  logout() {
+    localStorage.removeItem(this.tokenKey);
+    this.setAuthStatus(false);
+  }
+
+  private setAuthStatus(isAuthenticated: boolean): void {
+    this._authStatus.next(isAuthenticated);
   }
 
 }