+++ WorldCities/WorldCities/src/app/auth/login.component.ts
@@ -39,13 +39,15 @@ export class LoginComponent extends BaseFormComponent implements OnInit{
     loginRequest.email = this.form.controls['email'].value;
     loginRequest.password = this.form.controls["password"].value;
 
+
+
     this.authService
       .login(loginRequest)
       .subscribe(result => {
         console.log(result);
         this.loginResult = result;
-        if (result.success && result.token) {
-          localStorage.setItem(this.authService.tokenKey, result.token);
+        if (result.success) {
+          this.router.navigate(["/"]);
         }
       }, error => {
         console.log(error);